2019 - Filter revisions on username or display name * Add `MultiUsernameOrDisplayNameField` field * Replace field in `forms.py` with new fieldtype
I don't like how not-DRY this is when you look across the
"multi" fields created in the `form_fields` file.

I didn't want to update the `MultiUsernameField` class to
handle profile names because then it is not what it claims
to be, and might be a breaking change if used elsewhere.

I think we can: keep this, and move along OR create a new
base class these types can inherit from to knock out some
redundancy OR decide we don't care if `MultiUsernameField`
also checks for users based on profile name.

If we want to do more than what is in this PR, I'd suggest
it be another, separate effort.